.github/requirements.txt
openai>=1.0.0
PyGithub>=2.0.0
python-dotenv>=1.0.0
regex>=2023.0.0
requests>=2.25.0
PyGithub>=2.3.0